Move SettingsAware interface to the settings package
