+++ src/main/kotlin/dev/monosoul/jooq/settings/SettingsAware.kt
@@ -1,5 +1,6 @@
-package dev.monosoul.jooq
+package dev.monosoul.jooq.settings
 
+import dev.monosoul.jooq.callWith
 import dev.monosoul.jooq.settings.JooqDockerPluginSettings.WithContainer
 import dev.monosoul.jooq.settings.JooqDockerPluginSettings.WithoutContainer
 import groovy.lang.Closure
